requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
bzip2=1.0.6=h14c3975_1002
ca-certificates=2019.3.9=hecc5488_0
cairo=1.16.0=ha4e643d_1000
certifi=2019.3.9=py36_0
cloudpickle=1.0.0=py_0
cycler=0.10.0=py_1
cytoolz=0.9.0.1=py36h14c3975_1001
dask-core=1.2.2=py_0
dbus=1.13.6=he372182_0
decorator=4.4.0=py_0
expat=2.2.5=hf484d3e_1002
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.0=he983fc9_0
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.1.9=h516909a_0
glib=2.58.3=hf63aee3_1001
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=hf484d3e_1000
gst-plugins-base=1.14.4=hdf3bae2_1001
gstreamer=1.14.4=h66beb1c_1001
harfbuzz=2.4.0=h37c48d4_0
hdf5=1.10.4=nompi_h3c11f04_1106
icu=58.2=hf484d3e_1000
imageio=2.5.0=py36_0
jasper=1.900.1=h07fcdf6_1006
joblib=0.13.2=py_0
jpeg=9c=h14c3975_1001
kiwisolver=1.1.0=py36hc9558a2_0
lame=3.100=h14c3975_1001
libblas=3.8.0=10_openblas
libcblas=3.8.0=10_openblas
libffi=3.2.1=he1b5a44_1006
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=10_openblas
liblapacke=3.8.0=10_openblas
libpng=1.6.37=hed695b0_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h648cc4a_1001
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h576950b_1
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=h13577e0_0
matplotlib-base=3.1.0=py36hfd891ef_1
ncurses=6.1=hf484d3e_1002
nettle=3.4.1=h1bed415_1002
networkx=2.3=py_0
numpy=1.16.4=py36h95a1406_0
olefile=0.46=py_0
openblas=0.3.6=h6e990d7_2
opencv=4.1.0=py36h5517eff_4
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1b=h14c3975_1
pcre=8.41=hf484d3e_1003
pillow=6.0.0=py36he7afcd5_0
pip=19.1=py36_0
pixman=0.34.0=h14c3975_1003
pthread-stubs=0.4=h14c3975_1001
pyparsing=2.4.0=py_0
python=3.6.7=h381d211_1004
python-dateutil=2.8.0=py_0
pywavelets=1.0.3=py36hd352d35_1
qt=5.9.7=h52cfd70_2
readline=7.0=hf8c457e_1001
scikit-image=0.15.0=py36he1b5a44_0
scikit-learn=0.21.2=py36h627018c_0
scipy=1.3.0=py36h921218d_0
setuptools=41.0.1=py36_0
six=1.12.0=py36_1000
sqlite=3.28.0=h8b20d00_0
tk=8.6.9=h84994c4_1001
toolz=0.9.0=py_1
tornado=6.0.2=py36h516909a_0
wheel=0.33.4=py36_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.9=h516909a_1004
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.7=h14c3975_1000
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_1001
zlib=1.2.11=h14c3975_1004
